requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.2.1
Flask==3.1.2
# -e git+ssh://git@github.com/coghuynh/hunter.git@2150a395dda3a29a21fc46d0d0864577dc4aaf80#egg=hunter
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
neo4j==5.28.2
neomodel==5.5.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
pytz==2025.2
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
